adding spec tests

+++ visualization/spectrogram_window/spec_options.py
@@ -10,14 +10,14 @@ from scipy import signal
 
 class SpecOptions(QWidget):
     """ Class for spectrogram options window """
-    def __init__(self,data,parent):
+    def __init__(self, data, parent):
         """ Constructor """
         super().__init__()
         self.left = 10
         self.top = 10
         self.title = 'Select signal for spectrogram'
-        self.width = parent.width / 6
-        self.height = parent.height / 2.5
+        self.width = int(parent.width / 6)
+        self.height = int(parent.height / 2.5)
         self.data = data
         self.parent = parent
         self.setup_ui()
@@ -27,8 +27,8 @@ class SpecOptions(QWidget):
         self.setWindowTitle(self.title)
         center_point = QtWidgets.QDesktopWidget().availableGeometry().center()
 
-        self.setGeometry(center_point.x() - self.width / 2,
-            center_point.y() - self.height / 2, self.width, self.height)
+        self.setGeometry(center_point.x() - int(self.width / 2),
+            center_point.y() - int(self.height / 2), self.width, self.height)
 
         grid = QGridLayout()
 
@@ -36,34 +36,34 @@ class SpecOptions(QWidget):
         self.chn_combobox.addItems(["<select channel>"])
 
         lbl_info = QLabel("To hide the plot, click \n\"Clear\" and then \"Ok\".")
-        grid.addWidget(lbl_info,3,0)
+        grid.addWidget(lbl_info, 3, 0)
         my_font=QFont()
         my_font.setBold(True)
         lbl_info.setFont(my_font)
 
         lbl_chn = QLabel("Select a channel for \nspectrogram plotting: ")
-        grid.addWidget(lbl_chn,1,0)
-        grid.addWidget(self.chn_combobox,1,1,1,3)
+        grid.addWidget(lbl_chn, 1, 0)
+        grid.addWidget(self.chn_combobox, 1, 1, 1, 3)
 
         lblfsaxis = QLabel("x-axis (Hz): ")
-        grid.addWidget(lblfsaxis,2,0)
+        grid.addWidget(lblfsaxis, 2, 0)
         self.btn_get_min_fs = QDoubleSpinBox(self)
         self.btn_get_min_fs.setRange(0, self.parent.edf_info.fs / 2)
         self.btn_get_min_fs.setValue(self.data.min_fs)
         self.btn_get_min_fs.setDecimals(3)
-        grid.addWidget(self.btn_get_min_fs,2,1)
+        grid.addWidget(self.btn_get_min_fs, 2, 1)
         lblfsto = QLabel(" to ")
-        grid.addWidget(lblfsto,2,2)
+        grid.addWidget(lblfsto, 2, 2)
         self.btn_get_max_fs = QDoubleSpinBox(self)
         self.btn_get_max_fs.setDecimals(3)
         self.btn_get_max_fs.setRange(0, self.parent.edf_info.fs / 2)
         self.btn_get_max_fs.setValue(self.data.max_fs)
-        grid.addWidget(self.btn_get_max_fs,2,3)
+        grid.addWidget(self.btn_get_max_fs, 2, 3)
 
         self.btn_clear = QPushButton('Clear', self)
-        grid.addWidget(self.btn_clear,3,1,1,2)
+        grid.addWidget(self.btn_clear, 3, 1, 1, 2)
         self.btn_exit = QPushButton('Ok', self)
-        grid.addWidget(self.btn_exit,3,3)
+        grid.addWidget(self.btn_exit, 3, 3)
 
         self.setLayout(grid)
 
@@ -87,7 +87,7 @@ class SpecOptions(QWidget):
             self.close_window()
         else:
             for i in range(len(labels_to_plot) - 1):
-                self.labels_flipped.append(labels_to_plot[i+1])
+                self.labels_flipped.append(labels_to_plot[i + 1])
                 self.chn_combobox.addItems([labels_to_plot[len(labels_to_plot) - 1 - i]])
             if self.data.chn_plotted != -1:
                 self.chn_combobox.setCurrentIndex(len(labels_to_plot) - self.data.chn_plotted - 1)
